shuffle software around and remove fancy types
